Add files via upload
The main purpose of the program is to read QR codes using the webcam. To do this, we will simply press the "START READING BY CAMERA" button (which will display the camera viewfinder) and place the code to be read in front of the camera. In turn, we can also read QR codes, in "png" and "jpg" format stored on our computer (for which we will use the "LOAD FILE" option). Finally, we can also directly read a code that is visible in screen ("DETECT ON SCREEN" button).